Add render cropping to CLI (#535)

+++ TwitchDownloaderCLI/Modes/Arguments/ChatDownloadArgs.cs
@@ -17,10 +17,10 @@ namespace TwitchDownloaderCLI.Modes.Arguments
         public ChatCompression Compression { get; set; }
 
         [Option('b', "beginning", HelpText = "Time in seconds to crop beginning.")]
-        public int CropBeginningTime { get; set; }
+        public double CropBeginningTime { get; set; }
 
         [Option('e', "ending", HelpText = "Time in seconds to crop ending.")]
-        public int CropEndingTime { get; set; }
+        public double CropEndingTime { get; set; }
         
         [Option('E', "embed-images", Default = false, HelpText = "Embed first party emotes, badges, and cheermotes into the chat download for offline rendering.")]
         public bool EmbedData { get; set; }
